libmark6sg v1.1.0 with significantly improved read throughput (1-2 GB/s instead of 100 MB/s), new mk6copy copy/benchmark util, fuseMk6 new -v cmd line param
git-svn-id: https://svn.atnf.csiro.au/difx@6786 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/mark6sg/trunk/mark6sg/mark6_sg_utils.h
@@ -31,6 +31,7 @@
 
 #include <sys/types.h>
 #include <sys/stat.h>
+#include <stdint.h>
 
 // Defines
 #define MARK6_SG_SYNC_WORD          0xfeed6666
@@ -44,7 +45,7 @@ extern int m_m6sg_dbglevel;
 // Struct used by library to describe a group of files and their block content
 typedef struct m6sg_blockmeta_tt {
     int      file_id;        // index into the filename list (same list as passed to mark6_sg_blocklist())
-    int      blocknum;       // block number
+    int32_t  blocknum;       // block number
     int      packetsize;     // size of packets that Mark6 software thought were recorded
     off_t    file_offset;    // byte offset to the VLBI data in the file specified by file_id
     size_t   datalen;        // length in bytes of the VLBI data stored in the block
@@ -64,7 +65,7 @@ typedef struct m6sg_slistmeta_tt {
 extern int mark6_sg_filelist_from_name(const char* scanname, char*** filepathlist, char*** filenamelist);
 extern int mark6_sg_filelist_uniques(int nfiles, const char** filenamelist, char*** uniquenamelist);
 extern int mark6_sg_list_all_scans(char*** uniquenamelist);
-extern int mark6_sg_blocklist(int nfiles, const char** filenamelist, m6sg_blockmeta_t** blocklist);
+extern size_t mark6_sg_blocklist(int nfiles, const char** filenamelist, m6sg_blockmeta_t** blocklist);
 extern int mark6_sg_verbositylevel(int);
 extern int mark6_sg_collect_metadata(m6sg_slistmeta_t**);
